small log issue, getting rid of unused singleton classes
