+++ java-connectors/kafka-connect-azure-eventhubs/src/main/java/io/lenses/java/streamreactor/connect/azure/eventhubs/source/AzureConsumerRebalancerListener.java
@@ -41,7 +41,6 @@ public class AzureConsumerRebalancerListener implements ConsumerRebalanceListene
       partitionOffset.ifPresentOrElse(
           offset -> kafkaConsumer.seek(partition, offset.getOffsetValue()),
           () -> kafkaConsumer.seekToBeginning(Collections.singletonList(partition)));
-      log.info("Subscribed to topic: {}", partition.topic());
     });
   }
